navit: Add again Navit & all it's dependents
A work has been done to put the sofware up to date on GDP and make it
run on it. To make it work, need to add the old components that were
removed by commit acbe026 navit: Remove Navit & all it's dependents
on deprecated meta-genivi-dev. They will be updated separately.

[GDP-15]  Part of the Integrate LBS / FSA into GDP
[GDP-324] Navit not working on GDP10 Raspi3
[GDP-562] Part of Navigation build sometimes breaks

Signed-off-by: Matthias Bloch <matthias.bloch@mpsa.com>